--- errors.py
@@ -1,5 +1,0 @@
-class ValidationError(Exception):
-    pass
-
-class EmptyNotesError(Exception):
-    pass
