+++ src/errors/errors.py
@@ -1,2 +1,6 @@
 class ValidationError(Exception):
     pass
+
+
+class EmptyNotesError(Exception):
+    pass
