+++ src/models/note.py
@@ -45,6 +45,12 @@ class Note:
             if existing_tag == t:
                 self.tags[index] = Tag(new_tag.strip())
 
+    def has_in_content(self, line: str) -> bool:
+        return line.strip() in self.content.value
+
+    def has_in_title(self, line: str) -> bool:
+        return line.strip() in self.title.value
+
     def __hash__(self):
         return hash(self.title, self.content, tuple(self.tags))
